requirements.txt
requests>=2,<3
msgraphcore
azure-identity